# apps/backend/Models/EventType.cs
namespace BookingApi.Models;

public class EventType
{
    public Guid Id { get; set; }
    public string Name { get; set; } = string.Empty;
    public string Description { get; set; } = string.Empty;
    public int DurationMinutes { get; set; }
}